fix(embed): improve embed lookup loader
