--- datahub-web-react/src/app/embed/lookup/LookupLoading.tsx
@@ -1,8 +1,0 @@
-import { LoadingOutlined } from '@ant-design/icons';
-import styled from 'styled-components';
-
-const LookupLoading = styled(LoadingOutlined)`
-    font-size: 50px;
-`;
-
-export default LookupLoading;
