+++ datahub-web-react/src/app/embed/lookup/useGetEntityByUrl.ts
@@ -1,13 +1,16 @@
-import { useMemo } from 'react';
 import { useGetSearchResultsForMultipleQuery } from '../../../graphql/search.generated';
 import { FilterOperator } from '../../../types.generated';
 import { UnionType } from '../../search/utils/constants';
 import { generateOrFilters } from '../../search/utils/generateOrFilters';
+import { PageRoutes } from '../../../conf/Global';
+import { useEntityRegistry } from '../../useEntityRegistry';
+import { urlEncodeUrn } from '../../entity/shared/utils';
 
 const URL_FIELDS = ['externalUrl', 'chartUrl', 'dashboardUrl'] as const;
 
 const useGetEntityByUrl = (externalUrl: string) => {
-    const { data, loading, error } = useGetSearchResultsForMultipleQuery({
+    const registry = useEntityRegistry();
+    const { data, error } = useGetSearchResultsForMultipleQuery({
         variables: {
             input: {
                 query: '*',
@@ -25,20 +28,24 @@ const useGetEntityByUrl = (externalUrl: string) => {
         },
     });
 
-    const entities = data?.searchAcrossEntities?.searchResults.map((result) => result.entity) ?? [];
-    const count = entities.length;
-    const entity = count === 1 ? entities[0] : null;
+    const getLookupData = () => {
+        if (!data) return {} as const;
 
-    return useMemo(
-        () =>
-            ({
-                count,
-                entity,
-                loading,
-                error,
-            } as const),
-        [count, entity, error, loading],
-    );
+        const entities = data.searchAcrossEntities?.searchResults.map((result) => result.entity) ?? [];
+        const notFound = entities.length === 0;
+        const foundMultiple = entities.length > 1;
+        const entity = entities.length === 1 ? entities[0] : null;
+        const embedUrl = entity
+            ? `${PageRoutes.EMBED}/${registry.getPathName(entity.type)}/${urlEncodeUrn(entity.urn)}`
+            : null;
+
+        return { notFound, foundMultiple, embedUrl } as const;
+    };
+
+    return {
+        error,
+        ...getLookupData(),
+    } as const;
 };
 
 export default useGetEntityByUrl;
